Decoupling Description and Code views in a separate classes
